Add a simplistic PlaceholderService
The intent behind such a service is to harmonise the various templating systems to use a common bank of placeholders. PlaceholderParsers will be registered centrally using the Sponge registry system such that there is a standard way for all plugins to get a placeholder, alonge with a standard way to include them in text constructions.

This service is replaceable, allowing for a plugin to support a superset of the required base. Other plugins can also just make sure of the parser registry if they wish.

Notably, this PR does NOT include a templating system. This is intentional, to allow plugins to provide their own systems as now.

Part of the ideas from #1891